cli alterations
